+++ bot.py
@@ -4,10 +4,11 @@ import traceback
 import discord
 import discord.ext.commands as commands
 
-from cogs.admin import AdminCommands
+from cogs.admin import Admin
 from cogs.club_calendar import ClubCalendar
 from cogs.snipe import Snipes
 from cogs.soapbox import Soapbox
+from cogs.carepackage import CarePackage
 from data import api
 
 log = logging.getLogger(__name__)
@@ -20,8 +21,9 @@ class SnipeBot(commands.Bot):
 
         self.add_cog(Soapbox(self))
         self.add_cog(Snipes(self, day, start, end))
-        self.add_cog(AdminCommands(self))
+        self.add_cog(Admin(self))
         self.add_cog(ClubCalendar(self))
+        self.add_cog(CarePackage(self))
 
     async def on_command_error(self, ctx, error):
         if isinstance(error, commands.CommandNotFound):
