+++ cogs/snipe.py
@@ -173,68 +173,3 @@ class Snipes(commands.Cog):
         with open('whitelist', 'w') as f:
             for i in self.whitelist:
                 f.write(str(i) + '\n')
-
-    # region CarePackage Commands
-    @commands.command(name='smoke_bomb', hidden=True)
-    async def use_smoke_bomb(self, ctx: commands.Context):
-        if Database.has_smoke_bomb(ctx.author.id):
-            if Database.use_smoke_bomb(ctx.author.id):
-                await ctx.send(ctx.author.display_name + ' has used a smoke bomb and is now immune for the next 3 hours!')
-            else:
-                await ctx.send('Error using smoke bomb.')
-        else:
-            await ctx.send('You don\'t have a smoke bomb to use!')
-
-    @commands.command(name='set_carepackage', hidden=True)
-    @commands.has_role(item="Dev Team")
-    async def set_carepackage_cmd(self, ctx: commands.Context, keyword, time, hint):
-        await ctx.send(CarePackage.set_carepackage(keyword, time, hint))
-
-    @commands.command(name='get_hint', hidden=True)
-    async def get_carepackage_hint(self, ctx: commands.Context):
-        await ctx.send(CarePackage.get_hint())
-
-    @commands.command(name='get_rewards', hidden=True)
-    @commands.has_role(item="Dev Team")
-    async def get_carepackage_rewards(self, ctx: commands.Context):
-        rewards = Database.get_rewards()
-
-        sendingStr = ''
-
-        for reward in rewards:
-            sendingStr += reward[0] + '\n\t\u2192 ' + reward[1] + '\n'
-
-        await ctx.send('```' + sendingStr + '```')
-
-    @commands.command(name='announce_carepackage', hidden=True)
-    @commands.has_role(item="Dev Team")
-    async def announce_carepackage(self, ctx: commands.Context):
-        await self.snipe_channel.send('{} A carepackage is spawning soon!'.format(ctx.guild.default_role))
-
-    @commands.command(name='guess', hidden=True)
-    async def guess_keyword(self, ctx: commands.Context, keyword):
-        if Database.isRespawning(ctx.author.id):
-            await ctx.send('Sorry, you can\'t claim the carepackage if you\'re dead!')
-            return
-
-        success = CarePackage.isKeyword(keyword)
-
-        if not success:
-            await ctx.send('Sorry {}, that is not the keyword.'.format(ctx.author.display_name))
-            return
-
-        reward = Database.get_random_reward()
-        msg = Rewards.get_reward(reward[0], ctx.author.id)
-
-        await ctx.send('{} guessed the keyword correctly! You open the care package and earn {}!'.format(ctx.author.display_name, msg))
-
-    @commands.command(name='give_carepackage', hidden=True)
-    @commands.has_role(item='Dev Team')
-    async def give_carepackage(self, ctx: commands.Context, member):
-
-        reward = Database.get_random_reward()
-        msg = Rewards.get_reward(reward[0], ctx.author.id)
-
-        await ctx.send('{} opens the care package and earn {}!'.format(member.display_name, msg))
-
-    # endregion
